user delete confirm

+++ src/screens/UserListScreen/UserListScreen.jsx
@@ -30,7 +30,11 @@ const UserListScreen = () => {
     }, [dispatch, navigate, successDelete])
 
     const deleteHandler = (id) => {
-        dispatch(deleteUser(id))
+        if (window.confirm('Are you sure you want to delete this user?')) {
+            dispatch(deleteUser(id))
+        }
+
+        
     }
 
   return (
